requirements.txt
streamlit==1.33.0
supabase==2.4.0
reportlab==4.1.0
matplotlib==3.8.4
pillow==10.3.0
python-dotenv==1.0.1
streamlit-js-eval==0.1.7
